LPS-159234 Move internal/servlet/taglib/ui/constants to internal/constants
