+++ modules/apps/commerce/commerce-currency-web/src/main/java/com/liferay/commerce/currency/web/internal/frontend/taglib/servlet/taglib/CommerceCurrencyExchangeRateScreenNavigationCategory.java
@@ -19,8 +19,8 @@ import com.liferay.commerce.currency.model.CommerceCurrencyConstants;
 import com.liferay.commerce.currency.service.CommerceCurrencyService;
 import com.liferay.commerce.currency.util.CommercePriceFormatter;
 import com.liferay.commerce.currency.util.ExchangeRateProviderRegistry;
+import com.liferay.commerce.currency.web.internal.constants.CommerceCurrencyScreenNavigationConstants;
 import com.liferay.commerce.currency.web.internal.display.context.CommerceCurrenciesDisplayContext;
-import com.liferay.commerce.currency.web.internal.servlet.taglib.ui.constants.CommerceCurrencyScreenNavigationConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
 import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
